renaming files with structs
