+++ Source/LibraryClient.cpp
@@ -11,5 +11,9 @@
  * main
  ******************************************************************************/
 int main( int argc, char *argv[] ) {
+  std::cout << "usage: LibraryClient functionIdx sizes" << std::endl;
+  for (unsigned int i = 0; i < numFunctions; i++) {
+    std::cout << "(" << i << ") " << functionNames[i] << std::endl;
+  }
   return 0;
 }
